# backend/specs/swagger.yaml
openapi: 3.0.2
servers:
  - url: /api
info:
  description: |-
    This is the Updog API documentation based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [http://swagger.io](http://swagger.io).
    - [The Updog repository](https://github.com/SE701Team2/Updog)
  version: 1.0.11
  title: Updog API - OpenAPI 3.0
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: 'se701updog@gmail.com'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /users:
    get:
      tags:
        - User
      summary: Get all user handles
      description: ''
      operationId: getAllUserHandles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  usernames:
                    type: array
                    items:
                      type: string
        '400':
          description: Auth token not provided
        '401':
          description: Auth token invalid
        '500':
          description: Internal Server Error
    post:
      tags:
        - User
      summary: Create user
      description: ''
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: user1
                nickname:
                  type: string
                  example: IamUser
                email:
                  type: string
                  example: user1@gmail.com
                password:
                  type: string
                  example: userPassword
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email has already been taken
        '500':
          description: Internal Server Error
    put:
      tags:
        - User
      summary: Modify user
      description: This can only be done by the logged in user.
      operationId: modifyUser
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: user1
                nickname:
                  type: string
                  example: IamUser
                bio:
                  type: string
                  example: This is my bio
                profilePic:
                  type: string
                  example: null
                profileBanner:
                  type: string
                  example: null
      responses:
        '200':
          description: The profile has been updated
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '500':
          description: Internal Server Error
    delete:
      tags:
        - User
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The user has been deleted
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '500':
          description: Internal Server Error
...
  '/users/{username}/follow':
    get:
      tags:
        - User
      summary: Get users who are following or are followers of this username
      description: ''
      operationId: getFollow
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  following:
                    $ref: '#/components/schemas/User'
                  followers:
                    $ref: '#/components/schemas/User'
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '500':
          description: Internal Server Error
    post:
      tags:
        - User
      summary: Follow a user
      description: ''
      operationId: followUser
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Followers'
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '404':
          description: User not found
        '409':
          description: Already following this user
        '500':
          description: Internal Server Error
    delete:
      tags:
        - User
      summary: Unfollow a user
      description: ''
      operationId: unfollowUser
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Followers'
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '404':
          description: User not found
        '500':
          description: Internal Server Error
  '/users/{username}/activity':
    get:
      tags:
        - User
      summary: Get a users activity
      description: ''
      operationId: getUserActivity
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  activity:
                    type: string
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '404':
          description: User not found
        '500':
          description: Internal Server Error
  /feed:
    get:
      tags:
        - User
      summary: Get a users feed
      description: 'This is generated from users this user follows'
      operationId: getFeed
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  activity:
                    type: string
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '500':
          description: Internal Server Error
  /notifications:
    get:
      tags:
        - User
      summary: Get a users notifications
      description: ''
      operationId: getNotifications
      parameters:
        - name: username
          in: path
          description: 'The username of the user that needs to be fetched.'
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  from:
                    type: string
                  post:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  content:
                    type: string
        '400':
          description: Auth token not provided
        '401':
          description: Access token invalid
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    User:
      x-swagger-router-model: io.swagger.updog.model.User
      properties:
        username:
          type: string
          example: theUser
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: 12345
        nickname:
          type: string
        profilePic:
          type: string
        profileBanner:
          type: string
        bio:
          type: string
        joinedDate:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        followed:
          type: array
          xml:
            name: followed
            wrapped: true
          items:
            $ref: '#/components/schemas/Followers'
        follower:
          type: array
          xml:
            name: follower
            wrapped: true
          items:
            $ref: '#/components/schemas/Followers'
        likedPosts:
          type: array
          xml:
            name: likedPosts
            wrapped: true
          items:
            $ref: '#/components/schemas/LikedPost'
        sharedPosts:
          type: array
          xml:
            wrapped: true
            name: sharedPosts
          items:
            $ref: '#/components/schemas/SharedPost'
      xml:
        name: user
      type: object
...
    Followers:
      x-swagger-router-model: io.swagger.updog.model.followers
      properties:
        followedId:
          type: integer
        followerId:
          type: integer
      type: object
    LikedPost:
      x-swagger-router-model: io.swagger.updog.model.likedpost
      properties:
        postId:
          type: integer
        userId:
          type: integer
      type: object
    SharedPost:
      x-swagger-router-model: io.swagger.updog.model.sharedpost
      properties:
        postId:
          type: integer
        userId:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT